trigger hooks
Given that the before and after hooks have a max run time of 3 seconds
it may be best to run all the callbacks for a hook in "parallel" and
resolve after all of them have finished. It is currently a waterfall;
the next callback will not be processed until the previous one has
resolved (if it returns a promise).